Adding MQTT and improving the quickstart (#70)
* Add mqtt endpoints

* add simulator

* update zilla port

* update version and add restart policies

+++ examples/quickstart/setup.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-docker-compose down --remove-orphans
-docker-compose build
-docker-compose up -d
+set -e
+if [ -x "$(command -v docker)" ]; then
+    docker-compose down --remove-orphans
+    docker-compose build
+    docker-compose up -d
+else
+    echo "Docker is required to run this setup."
+fi
